refactor: Request & Response DTO 수정
