+++ src/main/java/racingcar/game/dto/GameRequestDTO.java
@@ -1,11 +1,11 @@
-package racingcar.game.model;
+package racingcar.game.dto;
 
-public class PlayRequest {
+public class GameRequestDTO {
     
     private final String names;
     private final int count;
     
-    public PlayRequest(final String names, final int count) {
+    public GameRequestDTO(final String names, final int count) {
         this.names = names;
         this.count = count;
     }
